workload/image_classification/src/combined/src/requirements.txt
numpy
Pillow==8.1.0
